requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0 # Driver asíncrono para SQLite
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1